added logger base
